# Instance: Optimise DB queries during instance creation