This attempts to reduce queries/transactions used when creating an instance in an attempt to better support instance migration on high latency networks.

Related to #9861

This reduces the operation time when used with a 40ms latency by about 3s.

Before:
```
t=2022-02-08T17:08:02+0000 lvl=eror msg="tomp instancesPost" duration=16.465716182s
t=2022-02-08T17:08:46+0000 lvl=eror msg="tomp instancesPost" duration=21.270856828s
t=2022-02-08T17:09:24+0000 lvl=eror msg="tomp instancesPost" duration=16.394522801s
```

After:
```
t=2022-02-08T17:10:59+0000 lvl=eror msg="tomp instancesPost" duration=13.822343883s
t=2022-02-08T17:11:30+0000 lvl=eror msg="tomp instancesPost" duration=18.717397529s
t=2022-02-08T17:11:55+0000 lvl=eror msg="tomp instancesPost" duration=13.731331391s
```

Unfortunately whilst it does strip a few seconds off, it does not yet get it under the 10s limit (after which the migration source gives up).