@tomponline 2022-02-08
One area that stands out as quite slow is the bridged NIC duplicate MAC check:

https://github.com/lxc/lxd/blob/master/lxd/device/nic_bridged.go#L308-L387